Resolve TODO's, Remove unused methods, and reposition CamelModel to resolve circular import
