+++ src/steamship/base/__init__.py
@@ -1,3 +1,4 @@
+from .base import CamelModel
 from .client import Client
 from .configuration import Configuration
 from .error import SteamshipError
@@ -19,4 +20,5 @@ __all__ = [
     "Task",
     "TaskComment",
     "TaskState",
+    "CamelModel",
 ]
