+++ src/steamship/base/client.py
@@ -11,13 +11,13 @@ import inflection
 from pydantic import BaseModel, PrivateAttr
 from requests import Session
 
-from steamship.base.configuration import CamelModel, Configuration
+from steamship.base.base import CamelModel, to_camel
+from steamship.base.configuration import Configuration
 from steamship.base.error import SteamshipError
 from steamship.base.mime_types import MimeTypes
 from steamship.base.request import Request
 from steamship.base.response import Response, Task
 from steamship.base.tasks import TaskState
-from steamship.base.utils import to_camel
 from steamship.utils.url import Verb, is_local
 
 _logger = logging.getLogger(__name__)
