+++ src/steamship/base/tasks.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, List, Optional, Set, Type, TypeVar
 
 from pydantic import BaseModel
 
+from steamship.base import CamelModel
 from steamship.base.base import IResponse
-from steamship.base.configuration import CamelModel
 from steamship.base.error import SteamshipError
 from steamship.base.metadata import metadata_to_str, str_to_metadata
 from steamship.base.request import IdentifierRequest, Request
