+++ src/steamship/data/app_instance.py
@@ -4,8 +4,7 @@ from typing import Any, Dict, Type
 
 from pydantic import BaseModel
 
-from steamship.base import Client, Request, Response
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel, Client, Request, Response
 from steamship.data.space import Space
 
 
@@ -83,12 +82,25 @@ class AppInstance(CamelModel):
             if space and space.data:
                 self.space_handle = space.data.handle
 
-    def get(self, path: str, **kwargs):
+    def call(self, verb: str, path: str, **kwargs):
         self.load_missing_vals()
         if path[0] == "/":
             path = path[1:]
-        return self.client.get(
-            f"/{self.space_handle or '_'}/{self.handle or '_'}/{path}",  # TODO (enias): Fix code duplication
+        self.client.call(
+            verb=verb,
+            operation=self._resolve_path(path),
+            payload=kwargs,
+            is_app_call=True,
+            app_owner=self.user_handle,
+            app_id=self.app_id,
+            app_instance_id=self.id,
+            space_id=self.space_id,
+        )
+
+    def get(self, path: str, **kwargs):
+        return self.call(
+            "GET",
+            self._resolve_path(path),
             payload=kwargs,
             app_call=True,
             app_owner=self.user_handle,
@@ -98,11 +110,9 @@ class AppInstance(CamelModel):
         )
 
     def post(self, path: str, **kwargs):
-        self.load_missing_vals()
-        if path[0] == "/":
-            path = path[1:]
-        return self.client.post(
-            f"/{self.space_handle or '_'}/{self.handle or '_'}/{path}",
+        return self.call(
+            "POST",
+            self._resolve_path(path),
             payload=kwargs,
             app_call=True,
             app_owner=self.user_handle,
@@ -110,6 +120,9 @@ class AppInstance(CamelModel):
             app_instance_id=self.id,
             space_id=self.space_id,
         )
+
+    def _resolve_path(self, path: str):
+        return f"/{self.space_handle or '_'}/{self.handle or '_'}/{path}"
 
     def full_url_for(self, path: str):
         return f"{self.invocation_url}{path}"
