+++ src/steamship/data/plugin_instance.py
@@ -4,8 +4,7 @@ from typing import Any, Dict, Optional, Type, Union
 
 from pydantic import BaseModel
 
-from steamship.base import Client, Request, Response
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel, Client, Request, Response
 from steamship.data.plugin import (
     HostingCpu,
     HostingEnvironment,
@@ -77,8 +76,7 @@ class PluginInstance(CamelModel):
         upsert: bool = True,
         config: Dict[str, Any] = None,
     ) -> Response[PluginInstance]:
-        # TODO (enias): Document
-        """Create a plugin instance
+        """Create a plugin instance.
 
         When handle is empty the engine will automatically assign one
         upsert controls whether we want to re-use an existing plugin instance or not."""
