+++ src/steamship/data/user.py
@@ -5,8 +5,7 @@ from typing import Any, Dict, Type
 from pydantic import BaseModel
 
 from steamship.app import Response
-from steamship.base import Client
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel, Client
 
 
 class User(CamelModel):
