+++ src/steamship/plugin/inputs/train_plugin_input.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from pydantic import Field
 
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel
 
 
 class TrainPluginInput(CamelModel):
