+++ src/steamship/plugin/inputs/training_parameter_plugin_input.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Dict, Optional
 
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel
 from steamship.plugin.inputs.export_plugin_input import ExportPluginInput
