+++ src/steamship/plugin/outputs/embedded_items_plugin_output.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List
 
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel
 
 
 class EmbeddedItemsPluginOutput(CamelModel):
