+++ src/steamship/plugin/outputs/raw_data_plugin_output.py
@@ -5,9 +5,8 @@ from typing import Any, Optional, Type, Union
 
 from pydantic import BaseModel
 
-from steamship.base import MimeTypes
+from steamship.base import CamelModel, MimeTypes
 from steamship.base.binary_utils import flexi_create
-from steamship.base.configuration import CamelModel
 
 
 class RawDataPluginOutput(CamelModel):
