+++ src/steamship/plugin/outputs/train_plugin_output.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from steamship.base.configuration import CamelModel
+from steamship.base import CamelModel
 
 
 class TrainPluginOutput(CamelModel):
