+++ tests/steamship_tests/base/test_binary_utils.py
@@ -1,8 +1,8 @@
 import base64
 import json
 
+from steamship.base import CamelModel
 from steamship.base.binary_utils import flexi_create
-from steamship.base.configuration import CamelModel
 from steamship.plugin.outputs.raw_data_plugin_output import RawDataPluginOutput
